chore: Move admin features to a dedicated endpoint and file
This will simplify things when configuring our reverse proxy. All requests to the admin endpoint will be rejected.
